refactor(api): restructure packages; remove food_products; add V4 Flyway to drop table
