+++ services/api/src/main/java/com/chubini/pku/PkuApiApplication.java
@@ -1,4 +1,4 @@
-package com.chubini.pku.api;
+package com.chubini.pku;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -9,3 +9,6 @@ public class PkuApiApplication {
         SpringApplication.run(PkuApiApplication.class, args);
     }
 }
+
+
+
