+++ services/api/src/main/java/com/chubini/pku/config/GlobalExceptionHandler.java
@@ -1,7 +1,4 @@
-package main.java.com.chubini.pku.config;
-
-import main.java.com.chubini.pku.products.ProductNotFoundException;
-import main.java.com.chubini.pku.products.ProductUploadException;
+package com.chubini.pku.config;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -14,30 +11,6 @@ import java.util.Map;
 
 @ControllerAdvice
 public class GlobalExceptionHandler {
-
-    @ExceptionHandler(ProductNotFoundException.class)
-    public ResponseEntity<Object> handleProductNotFound(ProductNotFoundException ex, WebRequest request) {
-        Map<String, Object> body = new HashMap<>();
-        body.put("timestamp", LocalDateTime.now());
-        body.put("status", HttpStatus.NOT_FOUND.value());
-        body.put("error", "Product Not Found");
-        body.put("message", ex.getMessage());
-        body.put("path", request.getDescription(false));
-        
-        return new ResponseEntity<>(body, HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler(ProductUploadException.class)
-    public ResponseEntity<Object> handleProductUpload(ProductUploadException ex, WebRequest request) {
-        Map<String, Object> body = new HashMap<>();
-        body.put("timestamp", LocalDateTime.now());
-        body.put("status", HttpStatus.BAD_REQUEST.value());
-        body.put("error", "Product Upload Error");
-        body.put("message", ex.getMessage());
-        body.put("path", request.getDescription(false));
-        
-        return new ResponseEntity<>(body, HttpStatus.BAD_REQUEST);
-    }
 
     @ExceptionHandler(Exception.class)
     public ResponseEntity<Object> handleGenericException(Exception ex, WebRequest request) {
