--- services/api/src/main/java/com/chubini/pku/products/ProductNotFoundException.java
@@ -1,12 +1,0 @@
-package main.java.com.chubini.pku.products;
-
-public class ProductNotFoundException extends RuntimeException {
-    
-    public ProductNotFoundException(String message) {
-        super(message);
-    }
-    
-    public ProductNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
