--- services/api/src/main/java/com/chubini/pku/products/ProductUploadException.java
@@ -1,12 +1,0 @@
-package main.java.com.chubini.pku.products;
-
-public class ProductUploadException extends RuntimeException {
-    
-    public ProductUploadException(String message) {
-        super(message);
-    }
-    
-    public ProductUploadException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
