+++ services/api/src/main/resources/db/migration/V4__drop_food_products.sql
@@ -1,0 +1,4 @@
+DROP TABLE IF EXISTS food_products CASCADE;
+
+
+
